+++ logging/promoter.go
@@ -1,9 +1,16 @@
 package logging
 
-// Promoter is an implementation of Logger that forwards all messages to a
+// Promote returns a Logger that forwards all messages to a target logger as
+// non-debug messages. Thus, it "promotes" debug messages to the non-debug
+// level.
+func Promote(target Logger) Logger {
+	return promoter{target}
+}
+
+// promoter is an implementation of Logger that forwards all messages to a
 // target logger as non-debug messages. Thus, it "promotes" debug messages to
 // the non-debug level.
-type Promoter struct {
+type promoter struct {
 	// Target is the logger to which messages are forwarded.
 	Target Logger
 }
@@ -15,7 +22,7 @@ type Promoter struct {
 // operating the application, such as the end-user or operations staff.
 //
 // f is the format specifier, as per fmt.Printf(), etc.
-func (l *Promoter) Log(f string, v ...interface{}) {
+func (l promoter) Log(f string, v ...interface{}) {
 	Log(l.Target, f, v...)
 }
 
@@ -23,7 +30,7 @@ func (l *Promoter) Log(f string, v ...interface{}) {
 //
 // It should be used for messages that are intended for people responsible for
 // operating the application, such as the end-user or operations staff.
-func (l *Promoter) LogString(s string) {
+func (l promoter) LogString(s string) {
 	LogString(l.Target, s)
 }
 
@@ -36,7 +43,7 @@ func (l *Promoter) LogString(s string) {
 // that maintain the application.
 //
 // f is the format specifier, as per fmt.Printf(), etc.
-func (l *Promoter) Debug(f string, v ...interface{}) {
+func (l promoter) Debug(f string, v ...interface{}) {
 	l.Log(f, v...)
 }
 
@@ -46,7 +53,7 @@ func (l *Promoter) Debug(f string, v ...interface{}) {
 //
 // It should be used for messages that are intended for the software developers
 // that maintain the application.
-func (l *Promoter) DebugString(s string) {
+func (l promoter) DebugString(s string) {
 	l.LogString(s)
 }
 
@@ -59,6 +66,6 @@ func (l *Promoter) DebugString(s string) {
 // calling IsDebug(), however it can be used to check if debug logging is
 // necessary before executing expensive code that is only used to obtain debug
 // information.
-func (l *Promoter) IsDebug() bool {
+func (l promoter) IsDebug() bool {
 	return true
 }